requirements.txt
aiocache==0.11.1
aiocqhttp==1.2.5
aiofiles==0.4.0
APScheduler==3.6.3
astroid==2.3.3
beautifulsoup4==4.8.2
blinker==1.4
brainyquote==0.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.3.31
httpx>=0.12.1
Hypercorn==0.9.4
hyperframe==5.2.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
networkx==2.4
nonebot==1.5.0
numpy==1.21.0
Pillow==9.0.0
priority==1.3.0
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
Quart==0.11.5
requests>=2.23.0
rfc3986==1.3.2
six==1.14.0
sniffio==1.1.0
soupsieve==2.0
toml==0.10.0
typing-extensions==3.7.4.2
tzlocal==2.0.0
urllib3==1.26.5
Werkzeug==1.0.1
wordcloud==1.6.0
wrapt==1.11.2
wsproto==0.15.0
